Fix Styles.AddRequiredResource method
when called multiple times, it would try to append to the
string[] property Styles.RequiredResources, which wasn't supported
by the default attribute merger.

I have split the DefaultAttributeMerger into abstract base
class which only implements the lookups and the Default
which now implements collection concatenation (for server
values, not for bindings)
